+++ include/FastLS/object/Path.hpp
@@ -13,20 +13,20 @@ class Path : public ObjectBase {
     cylinder_.SetColor(color_);
     sphere_.SetColor(color_);
     for (size_t i = 1; i < path_.size(); ++i) {
-      glm::vec3 p1 = path_[i - 1];
-      glm::vec3 p2 = path_[i];
-      glm::vec3 direction = glm::normalize(p2 - p1);
+      glm::dvec3 p1 = path_[i - 1];
+      glm::dvec3 p2 = path_[i];
+      glm::dvec3 direction = glm::normalize(p2 - p1);
       float length = glm::distance(p1, p2);
-      glm::vec3 mid_point = (p1 + p2) / 2.0F;
+      glm::dvec3 mid_point = (p1 + p2) / 2.0;
       cylinder_.SetPos(mid_point);
-      cylinder_.SetSize(glm::vec3(0.1F, 0.1F, length));
+      cylinder_.SetSize(glm::dvec3(0.1, 0.1, length));
       cylinder_.SetQuatRotation(
-          glm::quatLookAt(direction, glm::vec3(0.0F, 0.0F, 1.0F)));
+          glm::quatLookAt(direction, glm::dvec3(0.0, 0.0, 1.0)));
       cylinder_.UpdateMatrix();
       cylinder_.Draw(program);
 
       if (is_sphere_) {
-        sphere_.SetPos(p2).SetSize(glm::vec3(0.2F, 0.2F, 0.2F));
+        sphere_.SetPos(p2).SetSize(glm::dvec3(0.2, 0.2, 0.2));
         sphere_.UpdateMatrix();
         sphere_.Draw(program);
       }
